tests/requirements.txt
requests
python-dotenv
psutil